requirements.txt
boto3==1.18.56
pillow==8.3.2